refactor integration test (#1940)
