+++ shenyu-integrated-test/shenyu-integrated-test-http/src/main/java/org/apache/shenyu/integrated/test/http/HttpIntegratedBootstrap.java
@@ -15,22 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.integratedtest.http;
+package org.apache.shenyu.integrated.test.http;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
- * Hello world.
+ * The type Http integrated bootstrap.
  */
 @SpringBootApplication
-public class App {
-
+public class HttpIntegratedBootstrap {
+    
     /**
      * main method of App.
+     *
      * @param args args
      */
     public static void main(final String[] args) {
-        SpringApplication.run(App.class);
+        SpringApplication.run(HttpIntegratedBootstrap.class);
     }
 }
