+++ shenyu-integrated-test/shenyu-integrated-test-spring-cloud/src/main/java/org/apache/shenyu/integratedtest/springcloud/SpringCloudIntegratedBootstrap.java
@@ -17,22 +17,23 @@
 
 package org.apache.shenyu.integratedtest.springcloud;
 
-import org.apache.shenyu.integratedtest.common.AbstractTest;
-import org.apache.shenyu.integratedtest.common.dto.OrderDTO;
-import org.apache.shenyu.integratedtest.common.helper.HttpHelper;
-import org.junit.Test;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
-import java.io.IOException;
-
-import static org.junit.Assert.assertEquals;
-
-public class HelloWorldTest extends AbstractTest {
-
-    @Test
-    public void testHelloWorld() throws IOException {
-        OrderDTO order = new OrderDTO("123", "Phoenix");
-        order = HttpHelper.INSTANCE.postGateway("/springcloud/order/save", order, OrderDTO.class);
-        assertEquals("hello world spring cloud save order", order.getName());
+/**
+ * The type Spring cloud integrated bootstrap.
+ */
+@SpringBootApplication
+@EnableDiscoveryClient
+public class SpringCloudIntegratedBootstrap {
+    
+    /**
+     * main method of App.
+     *
+     * @param args args
+     */
+    public static void main(final String[] args) {
+        SpringApplication.run(SpringCloudIntegratedBootstrap.class);
     }
-
 }
